Use QThread instead of Thread
